Remove reference to Elagabalus
Roman emperor Elagabalus is a poor historical representation for the trans community. Wikipedia here: https://en.wikipedia.org/wiki/Elagabalus

1. If Elagabalus was indeed trans, then the wikipedia article is misgendering
2. "Elagabalus developed a reputation among his contemporaries for extreme eccentricity, decadence, zealotry, and sexual promiscuity. This tradition has persisted, and among writers of the early modern age he suffered one of the worst reputations among Roman emperors"
3. "According to Barthold Georg Niebuhr, "the name Elagabalus is branded in history above all others" because of his "unspeakably disgusting life".[9] An example of a modern historian's assessment is Adrian Goldsworthy's: "Elagabalus was not a tyrant, but he was an incompetent, probably the least able emperor Rome had ever had."[10] "

While we can debate about if history is prejudice against Elagabalus (it most certainly is), we can all see that she is framed in such a negative light that using her to show trans people have been around for a long time is not helpful.