Introduction edits
After speaking with a non-binary person and another trans person who is a historian, and based on my own experience as a trans woman, I wanted to suggest some edits to the introduction. These edits adhere more closely to the quoted definition of gender, reduce references to physical anatomy, focus on broader historical claims which are unassailably well-sourced, and strive to foster positive reactions within the larger trans community.